Make _ShowSuggestions work in web components
The package can mostly work inside a web component as long as you
call getOrCreateInstance() and pass it the elements directly rather
than passing selectors to init().

However, when it attempts to show the suggestions you get nothing
because activeElement only gives you the active component, not the
individual element within it.  So use a little trick found on the
interwebs to drill recursively into the shadow DOM if necessary,
but it starts from document.activeElement so if you have only light
DOM, it does the same thing it always did.